small cleanup
git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@1222 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ pintable.cpp
@@ -543,7 +543,7 @@ STDMETHODIMP ScriptGlobalTable::put_DMDPixels(VARIANT pVal) //!! use 64bit inste
 
 	if(psa && g_pplayer && g_pplayer->m_dmdx > 0 && g_pplayer->m_dmdy > 0)
 	{
-		const unsigned int size = g_pplayer->m_dmdx*g_pplayer->m_dmdy;
+		const LONG size = g_pplayer->m_dmdx*g_pplayer->m_dmdy;
 		if(g_pplayer->m_rawdmd.size() != size)
 		{
 			g_pplayer->m_rawdmd.resize(size);
